fix minimal build issue
